pyproject.toml
[project]
name = "myapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorlog>=6.9.0",
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "drf-yasg>=1.21.10",
    "faker>=37.6.0",
    "psycopg2-binary>=2.9.10",
    "pygments>=2.19.2",
    "python-dotenv>=1.1.1",
]
